Started on the "runtime" module
The idea of the runtime module is to recieve a collection of action
queues from wherever (monster queues, player queues, etc) and decide
what order the moves should be made in.

The runtime also introduces a run loop which means that there is
finally a disconnect between when the players send their updates and
when the server decides to update the gamestate. (Previously, we'd
send updates when the server recieved instructions from players only).

The runtime will also trigger the gamestate changes (moving entities
around) and actually initiate the messages to be sent to the players.

Tl;dr: big steps forward. If you have two players with a queue each,
they actually take turns moving now. Though it's pretty much round-robin
atm.

Signed-off-by: Andrei Thorp <garoth@gmail.com>