Added general entity reposition hook (for boulder)
An entity may now call a function when it gets repositioned (i.e.
changes location). This is used for the rolling boulder to make it
"maintain momentum". Basically, every time it repositions, it calculates
the primary direction of movement, and then keeps going that way.

This means that as it's just normally rolling along, it's re-asserting
it's direction constantly, but stays rolling in the same direction.
However, if it gets pushed by something, then it will change its
direction of movement.

Tl;dr you can now redirect boulders at people using push.

Signed-off-by: Andrei Thorp <garoth@gmail.com>