Created free-standing server library in JS
This separates our direct manipulation of the websocket that used to be
in main.js into its own file, server.js. This provides a more convenient
library for interfacing with the server's commands, and automates
parsing of the json and stuff. Also updated the existing client to use
this new lib.

Signed-off-by: Andrei Thorp <athorp@evernote.com>